requirements.txt
tdqm
torch
torchvision
transformers
numpy
pandas
openai
zhipuai
zai-sdk
accelerate
sentencepiece